Data structure changes
Change the structure of the SavedVariable data structure to make counts
possible. Storing frequent toggle type events as log entries would not
be feasible.

These changes brought some issues to the forefront and they had to be
addressed if the addon would continue working. Notably was the
ADDON_LOADED event handling which hadcaused the SavedVariables to not be
handled corrrectly on load.