Add heartbeats
Added heartbeat functionality to log location and played time every 30
and 900 seconds respectively. Played can likely be longer, but might be
nescessary to keep played clock in sync with local clock. Can be removed
if not needed. To accomplish this the event  handlers were pulled out
and made named functions instead of anonymous ones created in the
SetScript call. This allowed the location heartbeat to simply call the
eventHandler with a fake HEARTBEAT_LOCATION event.