Spell Counts!
Addon now counts spell stats. Tracks casts, success, afailures, and
interrupts (sorta). This is done in a new bucket system instead of logs
so that the storage space is based on the number of different spells,
not the number of casts as it would have been with logs. So if you cast
only 5 spells ever, it's the same storage if you cast them all once, or
a billion times.

Will also be adding durations for things like time spent moving, time on
a mount, time casting, time eating, time in zones, time on boats, taxis,
quest vehicles, and so on. This should all be low storage as well and I
think give a huge amount of info about how your character lives its
life. Course, I'm prolly wrong.